+++ lib/glimmer/wx/control_proxy/slider_proxy.rb
@@ -1,4 +1,4 @@
-# Copyright (c) 2023-2024 Andy Maleh
+# Copyright (c) 2021-2024 Andy Maleh
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -19,16 +19,37 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/dsl/static_expression'
+require 'glimmer/wx/control_proxy'
 
 module Glimmer
-  module DSL
-    module Wx
-      class AboutBoxExpression < StaticExpression
-        def interpret(parent, keyword, *args, &block)
-          ::Wx.about_box(*args)
+  module Wx
+    class ControlProxy
+      # Proxy for Wx slider double objects
+      #
+      # Follows the Proxy Design Pattern
+      class SliderProxy < ControlProxy
+        include ::Wx::IDHelper
+        
+        class << self
+          def new_control(keyword, parent, args)
+            args = args.clone || []
+            if args.last.is_a?(Hash)
+              args[-1] = args[-1].clone
+            end
+            value = 0
+            min = 0
+            max = 100
+            style = 0
+            ::Wx.const_get(wx_constant_symbol(keyword)).new(parent, next_id, value, min, max, style: style)
+          end
+        end
+      
+        def data_bind_write(property, model_binding)
+          handle_listener('on_slider') { model_binding.call(value) } if property == 'value'
         end
       end
     end
   end
 end
+
+Dir[File.expand_path("./#{File.basename(__FILE__, '.rb')}/*.rb", __dir__)].each {|f| require f}